refactor: clean up workflow builder exports and remove unused workflow_screens file
