+++ packages/api_testing_suite/lib/src/tests/workflow_builder_test.dart
@@ -1,5 +1,4 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:api_testing_suite/src/workflow_builder/workflow_node.dart';
 
 void main() {
   group('Workflow Builder Tests', () {
