--- packages/api_testing_suite/lib/src/workflow_builder/workflow_screens.dart
@@ -1,31 +1,0 @@
-// Main components
-export 'workflow_builder.dart';
-export 'components/workflow_builder_screen.dart';
-export 'execution/workflow_executor.dart';
-
-// Providers
-export 'providers/providers.dart';
-
-// Models
-export 'models/models.dart';
-
-// Widgets
-export 'widgets/widgets.dart';
-
-// Common widgets with hide
-export '../common/widgets/widgets.dart';
-
-// Screens entry points
-import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'components/workflow_builder_screen.dart';
-
-/// A wrapper widget that provides an entry point to the workflow builder
-class WorkflowBuilderEntry extends ConsumerWidget {
-  const WorkflowBuilderEntry({super.key});
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    return const WorkflowBuilderScreen();
-  }
-}
